LogPoller - support downstream testing (#8648)
