+++ core/chains/evm/logpoller/log_poller_test.go
@@ -1,22 +1,20 @@
-package logpoller
+package logpoller_test
 
 import (
 	"context"
-	"database/sql"
 	"fmt"
 	"math/big"
-	"strings"
 	"testing"
 	"time"
 
-	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind/backends"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/rawdb"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/jackc/pgconn"
+	"github.com/ethereum/go-ethereum/params"
 	"github.com/leanovate/gopter"
 	"github.com/leanovate/gopter/gen"
 	"github.com/leanovate/gopter/prop"
@@ -24,11 +22,11 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/tendermint/tendermint/libs/rand"
-	"go.uber.org/zap/zapcore"
-	"go.uber.org/zap/zaptest/observer"
 
 	"github.com/smartcontractkit/chainlink/core/chains/evm/client"
+	"github.com/smartcontractkit/chainlink/core/chains/evm/logpoller"
 	"github.com/smartcontractkit/chainlink/core/gethwrappers/generated/log_emitter"
+	"github.com/smartcontractkit/chainlink/core/internal/cltest/heavyweight"
 	"github.com/smartcontractkit/chainlink/core/internal/testutils"
 	"github.com/smartcontractkit/chainlink/core/internal/testutils/pgtest"
 	"github.com/smartcontractkit/chainlink/core/logger"
@@ -36,56 +34,345 @@ import (
 	"github.com/smartcontractkit/chainlink/core/utils"
 )
 
-var (
-	EmitterABI, _ = abi.JSON(strings.NewReader(log_emitter.LogEmitterABI))
-)
-
-func GenLog(chainID *big.Int, logIndex int64, blockNum int64, blockHash string, topic1 []byte, address common.Address) Log {
-	return Log{
-		EvmChainId:  utils.NewBig(chainID),
-		LogIndex:    logIndex,
-		BlockHash:   common.HexToHash(blockHash),
-		BlockNumber: blockNum,
-		EventSig:    common.BytesToHash(topic1),
-		Topics:      [][]byte{topic1},
-		Address:     address,
-		TxHash:      common.HexToHash("0x1234"),
-		Data:        append([]byte("hello "), byte(blockNum)),
-	}
+func logRuntime(t *testing.T, start time.Time) {
+	t.Log("runtime", time.Since(start))
 }
 
-func assertDontHave(t *testing.T, start, end int, orm *ORM) {
-	for i := start; i < end; i++ {
-		_, err := orm.SelectBlockByNumber(int64(i))
-		assert.True(t, errors.Is(err, sql.ErrNoRows))
-	}
-}
-
-func assertHaveCanonical(t *testing.T, start, end int, ec *backends.SimulatedBackend, orm *ORM) {
-	for i := start; i < end; i++ {
-		blk, err := orm.SelectBlockByNumber(int64(i))
-		require.NoError(t, err, "block %v", i)
-		chainBlk, err := ec.BlockByNumber(testutils.Context(t), big.NewInt(int64(i)))
-		require.NoError(t, err)
-		assert.Equal(t, chainBlk.Hash().Bytes(), blk.BlockHash.Bytes(), "block %v", i)
-	}
-}
-
-func TestLogPoller_Batching(t *testing.T) {
-	t.Parallel()
-	th := SetupTH(t, 2, 3, 2)
-	var logs []Log
-	// Inserts are limited to 65535 parameters. A log being 10 parameters this results in
-	// a maximum of 6553 log inserts per tx. As inserting more than 6553 would result in
-	// an error without batching, this test makes sure batching is enabled.
-	for i := 0; i < 15000; i++ {
-		logs = append(logs, GenLog(th.ChainID, int64(i+1), 1, "0x3", EmitterABI.Events["Log1"].ID.Bytes(), th.EmitterAddress1))
-	}
-	require.NoError(t, th.ORM.InsertLogs(logs))
-	lgs, err := th.ORM.SelectLogsByBlockRange(1, 1)
+func TestPopulateLoadedDB(t *testing.T) {
+	t.Skip("only for local load testing and query analysis")
+	lggr := logger.TestLogger(t)
+	_, db := heavyweight.FullTestDBV2(t, "logs_scale", nil)
+	chainID := big.NewInt(137)
+	_, err := db.Exec(`INSERT INTO evm_chains (id, created_at, updated_at) VALUES ($1, NOW(), NOW())`, utils.NewBig(chainID))
 	require.NoError(t, err)
-	// Make sure all logs are inserted
-	require.Equal(t, len(logs), len(lgs))
+	o := logpoller.NewORM(big.NewInt(137), db, lggr, pgtest.NewQConfig(true))
+	event1 := EmitterABI.Events["Log1"].ID
+	address1 := common.HexToAddress("0x2ab9a2Dc53736b361b72d900CdF9F78F9406fbbb")
+	address2 := common.HexToAddress("0x6E225058950f237371261C985Db6bDe26df2200E")
+
+	// We start at 1 just so block number > 0
+	for j := 1; j < 1000; j++ {
+		var logs []logpoller.Log
+		// Max we can insert per batch
+		for i := 0; i < 1000; i++ {
+			addr := address1
+			if (i+(1000*j))%2 == 0 {
+				addr = address2
+			}
+			logs = append(logs, logpoller.Log{
+				EvmChainId:  utils.NewBig(chainID),
+				LogIndex:    1,
+				BlockHash:   common.HexToHash(fmt.Sprintf("0x%d", i+(1000*j))),
+				BlockNumber: int64(i + (1000 * j)),
+				EventSig:    event1,
+				Topics:      [][]byte{event1[:], logpoller.EvmWord(uint64(i + 1000*j)).Bytes()},
+				Address:     addr,
+				TxHash:      common.HexToHash("0x1234"),
+				Data:        logpoller.EvmWord(uint64(i + 1000*j)).Bytes(),
+			})
+		}
+		require.NoError(t, o.InsertLogs(logs))
+	}
+	func() {
+		defer logRuntime(t, time.Now())
+		_, err := o.SelectLogsByBlockRangeFilter(750000, 800000, address1, event1)
+		require.NoError(t, err)
+	}()
+	func() {
+		defer logRuntime(t, time.Now())
+		_, err = o.SelectLatestLogEventSigsAddrsWithConfs(0, []common.Address{address1}, []common.Hash{event1}, 0)
+		require.NoError(t, err)
+	}()
+
+	// Confirm all the logs.
+	require.NoError(t, o.InsertBlock(common.HexToHash("0x10"), 1000000, time.Now()))
+	func() {
+		defer logRuntime(t, time.Now())
+		lgs, err := o.SelectDataWordRange(address1, event1, 0, logpoller.EvmWord(500000), logpoller.EvmWord(500020), 0)
+		require.NoError(t, err)
+		// 10 since every other log is for address1
+		assert.Equal(t, 10, len(lgs))
+	}()
+
+	func() {
+		defer logRuntime(t, time.Now())
+		lgs, err := o.SelectIndexedLogs(address2, event1, 1, []common.Hash{logpoller.EvmWord(500000), logpoller.EvmWord(500020)}, 0)
+		require.NoError(t, err)
+		assert.Equal(t, 2, len(lgs))
+	}()
+
+	func() {
+		defer logRuntime(t, time.Now())
+		lgs, err := o.SelectIndexLogsTopicRange(address1, event1, 1, logpoller.EvmWord(500000), logpoller.EvmWord(500020), 0)
+		require.NoError(t, err)
+		assert.Equal(t, 10, len(lgs))
+	}()
+}
+
+func TestLogPoller_Integration(t *testing.T) {
+	th := SetupTH(t, 2, 3, 2)
+	th.Client.Commit() // Block 2. Ensure we have finality number of blocks
+
+	err := th.LogPoller.RegisterFilter(logpoller.Filter{"Integration test", []common.Hash{EmitterABI.Events["Log1"].ID}, []common.Address{th.EmitterAddress1}})
+	require.NoError(t, err)
+	require.Len(t, th.LogPoller.Filter(nil, nil, nil).Addresses, 1)
+	require.Len(t, th.LogPoller.Filter(nil, nil, nil).Topics, 1)
+
+	// Calling Start() after RegisterFilter() simulates a node restart after job creation, should reload Filter from db.
+	require.NoError(t, th.LogPoller.Start(testutils.Context(t)))
+	require.Len(t, th.LogPoller.Filter(nil, nil, nil).Addresses, 1)
+	require.Len(t, th.LogPoller.Filter(nil, nil, nil).Topics, 1)
+
+	// Emit some logs in blocks 3->7.
+	for i := 0; i < 5; i++ {
+		_, err := th.Emitter1.EmitLog1(th.Owner, []*big.Int{big.NewInt(int64(i))})
+		require.NoError(t, err)
+		_, err = th.Emitter1.EmitLog2(th.Owner, []*big.Int{big.NewInt(int64(i))})
+		require.NoError(t, err)
+		th.Client.Commit()
+	}
+	// The poller starts on a new chain at latest-finality (5 in this case),
+	// replay to ensure we get all the logs.
+	require.NoError(t, th.LogPoller.Replay(testutils.Context(t), 1))
+
+	// We should immediately have all those Log1 logs.
+	logs, err := th.LogPoller.Logs(2, 7, EmitterABI.Events["Log1"].ID, th.EmitterAddress1)
+	require.NoError(t, err)
+	assert.Equal(t, 5, len(logs))
+	// Now let's update the Filter and replay to get Log2 logs.
+	err = th.LogPoller.RegisterFilter(logpoller.Filter{
+		"Emitter - log2", []common.Hash{EmitterABI.Events["Log2"].ID},
+		[]common.Address{th.EmitterAddress1},
+	})
+	require.NoError(t, err)
+	// Replay an invalid block should error
+	assert.Error(t, th.LogPoller.Replay(testutils.Context(t), 0))
+	assert.Error(t, th.LogPoller.Replay(testutils.Context(t), 20))
+	// Replay only from block 4, so we should see logs in block 4,5,6,7 (4 logs)
+	require.NoError(t, th.LogPoller.Replay(testutils.Context(t), 4))
+
+	// We should immediately see 4 logs2 logs.
+	logs, err = th.LogPoller.Logs(2, 7, EmitterABI.Events["Log2"].ID, th.EmitterAddress1)
+	require.NoError(t, err)
+	assert.Equal(t, 4, len(logs))
+
+	// Cancelling a replay should return an error synchronously.
+	ctx, cancel := context.WithCancel(context.Background())
+	cancel()
+	assert.True(t, errors.Is(th.LogPoller.Replay(ctx, 4), logpoller.ErrReplayAbortedByClient))
+
+	require.NoError(t, th.LogPoller.Close())
+}
+
+// Simulate a badly behaving rpc server, where unfinalized blocks can return different logs
+// for the same block hash.  We should be able to handle this without missing any logs, as
+// long as the logs returned for finalized blocks are consistent.
+func Test_BackupLogPoller(t *testing.T) {
+	th := SetupTH(t, 2, 3, 2)
+	// later, we will need at least 32 blocks filled with logs for cache invalidation
+	for i := int64(0); i < 32; i++ {
+		// to invalidate geth's internal read-cache, a matching log must be found in the bloom Filter
+		// for each of the 32 blocks
+		tx, err := th.Emitter1.EmitLog1(th.Owner, []*big.Int{big.NewInt(i + 7)})
+		require.NoError(t, err)
+		require.NotNil(t, tx)
+		th.Client.Commit()
+	}
+
+	ctx := testutils.Context(t)
+
+	filter1 := logpoller.Filter{"filter1", []common.Hash{
+		EmitterABI.Events["Log1"].ID,
+		EmitterABI.Events["Log2"].ID},
+		[]common.Address{th.EmitterAddress1}}
+	err := th.LogPoller.RegisterFilter(filter1)
+	require.NoError(t, err)
+
+	filters, err := th.ORM.LoadFilters(pg.WithParentCtx(testutils.Context(t)))
+	require.NoError(t, err)
+	require.Equal(t, 1, len(filters))
+	require.Equal(t, filter1, filters["filter1"])
+
+	err = th.LogPoller.RegisterFilter(
+		logpoller.Filter{"filter2",
+			[]common.Hash{EmitterABI.Events["Log1"].ID},
+			[]common.Address{th.EmitterAddress2}})
+	require.NoError(t, err)
+
+	defer th.LogPoller.UnregisterFilter("filter1")
+	defer th.LogPoller.UnregisterFilter("filter2")
+
+	// generate some tx's with logs
+	tx1, err := th.Emitter1.EmitLog1(th.Owner, []*big.Int{big.NewInt(1)})
+	require.NoError(t, err)
+	require.NotNil(t, tx1)
+
+	tx2, err := th.Emitter1.EmitLog2(th.Owner, []*big.Int{big.NewInt(2)})
+	require.NoError(t, err)
+	require.NotNil(t, tx2)
+
+	tx3, err := th.Emitter2.EmitLog1(th.Owner, []*big.Int{big.NewInt(3)})
+	require.NoError(t, err)
+	require.NotNil(t, tx3)
+
+	th.Client.Commit() // commit block 34 with 3 tx's included
+
+	h := th.Client.Blockchain().CurrentHeader() // get latest header
+	require.Equal(t, uint64(34), h.Number.Uint64())
+
+	// save these 3 receipts for later
+	receipts := rawdb.ReadReceipts(th.EthDB, h.Hash(), h.Number.Uint64(), params.AllEthashProtocolChanges)
+	require.NotZero(t, receipts.Len())
+
+	// Simulate a situation where the rpc server has a block, but no logs available for it yet
+	//  this can't happen with geth itself, but can with other clients.
+	rawdb.WriteReceipts(th.EthDB, h.Hash(), h.Number.Uint64(), types.Receipts{}) // wipes out all logs for block 34
+
+	body := rawdb.ReadBody(th.EthDB, h.Hash(), h.Number.Uint64())
+	require.Equal(t, 3, len(body.Transactions))
+	txs := body.Transactions                 // save transactions for later
+	body.Transactions = types.Transactions{} // number of tx's must match # of logs for GetLogs() to succeed
+	rawdb.WriteBody(th.EthDB, h.Hash(), h.Number.Uint64(), body)
+
+	currentBlock := th.PollAndSaveLogs(ctx, 1)
+	assert.Equal(t, int64(35), currentBlock)
+
+	// simulate logs becoming available
+	rawdb.WriteReceipts(th.EthDB, h.Hash(), h.Number.Uint64(), receipts)
+	require.True(t, rawdb.HasReceipts(th.EthDB, h.Hash(), h.Number.Uint64()))
+	body.Transactions = txs
+	rawdb.WriteBody(th.EthDB, h.Hash(), h.Number.Uint64(), body)
+
+	// flush out cached block 34 by reading logs from first 32 blocks
+	query := ethereum.FilterQuery{
+		FromBlock: big.NewInt(int64(2)),
+		ToBlock:   big.NewInt(int64(33)),
+		Addresses: []common.Address{th.EmitterAddress1},
+		Topics:    [][]common.Hash{{EmitterABI.Events["Log1"].ID}},
+	}
+	fLogs, err := th.Client.FilterLogs(ctx, query)
+	require.NoError(t, err)
+	require.Equal(t, 32, len(fLogs))
+
+	// logs shouldn't show up yet
+	logs, err := th.LogPoller.Logs(34, 34, EmitterABI.Events["Log1"].ID, th.EmitterAddress1)
+	require.NoError(t, err)
+	assert.Equal(t, 0, len(logs))
+
+	th.Client.Commit()
+	th.Client.Commit()
+
+	// Run ordinary poller + backup poller at least once
+	currentBlock, _ = th.LogPoller.LatestBlock()
+	th.LogPoller.PollAndSaveLogs(ctx, currentBlock+1)
+	th.LogPoller.BackupPollAndSaveLogs(ctx, 100)
+	currentBlock, _ = th.LogPoller.LatestBlock()
+
+	require.Equal(t, int64(37), currentBlock+1)
+
+	// logs still shouldn't show up, because we don't want to backfill the last finalized log
+	//  to help with reorg detection
+	logs, err = th.LogPoller.Logs(34, 34, EmitterABI.Events["Log1"].ID, th.EmitterAddress1)
+	require.NoError(t, err)
+	assert.Equal(t, 0, len(logs))
+
+	th.Client.Commit()
+
+	// Run ordinary poller + backup poller at least once more
+	th.LogPoller.PollAndSaveLogs(ctx, currentBlock+1)
+	th.LogPoller.BackupPollAndSaveLogs(ctx, 100)
+	currentBlock, _ = th.LogPoller.LatestBlock()
+
+	require.Equal(t, int64(38), currentBlock+1)
+
+	// all 3 logs in block 34 should show up now, thanks to backup logger
+	logs, err = th.LogPoller.Logs(30, 37, EmitterABI.Events["Log1"].ID, th.EmitterAddress1)
+	require.NoError(t, err)
+	assert.Equal(t, 5, len(logs))
+	logs, err = th.LogPoller.Logs(34, 34, EmitterABI.Events["Log2"].ID, th.EmitterAddress1)
+	require.NoError(t, err)
+	assert.Equal(t, 1, len(logs))
+	logs, err = th.LogPoller.Logs(32, 36, EmitterABI.Events["Log1"].ID, th.EmitterAddress2)
+	require.NoError(t, err)
+	assert.Equal(t, 1, len(logs))
+}
+
+func TestLogPoller_BlockTimestamps(t *testing.T) {
+	t.Parallel()
+	ctx := testutils.Context(t)
+	th := SetupTH(t, 2, 3, 2)
+
+	addresses := []common.Address{th.EmitterAddress1, th.EmitterAddress2}
+	topics := []common.Hash{EmitterABI.Events["Log1"].ID, EmitterABI.Events["Log2"].ID}
+
+	err := th.LogPoller.RegisterFilter(logpoller.Filter{"convertLogs", topics, addresses})
+	require.NoError(t, err)
+
+	blk, err := th.Client.BlockByNumber(ctx, nil)
+	require.NoError(t, err)
+	require.Equal(t, big.NewInt(1), blk.Number())
+	start := blk.Time()
+
+	// There is automatically a 10s delay between each block.  To make sure it's including the correct block timestamps,
+	// we introduce irregularities by inserting two additional block delays. We can't control the block times for
+	// blocks produced by the log emitter, but we can adjust the time on empty blocks in between.  Simulated time
+	// sequence:  [ #1 ] ..(10s + delay1).. [ #2 ] ..10s.. [ #3 (LOG1) ] ..(10s + delay2).. [ #4 ] ..10s.. [ #5 (LOG2) ]
+	const delay1 = 589
+	const delay2 = 643
+	time1 := start + 20 + delay1
+	time2 := time1 + 20 + delay2
+
+	require.NoError(t, th.Client.AdjustTime(delay1*time.Second))
+	hash := th.Client.Commit()
+
+	blk, err = th.Client.BlockByHash(ctx, hash)
+	require.NoError(t, err)
+	require.Equal(t, big.NewInt(2), blk.Number())
+	assert.Equal(t, time1-10, blk.Time())
+
+	_, err = th.Emitter1.EmitLog1(th.Owner, []*big.Int{big.NewInt(1)})
+	require.NoError(t, err)
+	hash = th.Client.Commit()
+
+	blk, err = th.Client.BlockByHash(ctx, hash)
+	require.NoError(t, err)
+	require.Equal(t, big.NewInt(3), blk.Number())
+	assert.Equal(t, time1, blk.Time())
+
+	require.NoError(t, th.Client.AdjustTime(delay2*time.Second))
+	th.Client.Commit()
+	_, err = th.Emitter2.EmitLog2(th.Owner, []*big.Int{big.NewInt(2)})
+	require.NoError(t, err)
+	hash = th.Client.Commit()
+
+	blk, err = th.Client.BlockByHash(ctx, hash)
+	require.NoError(t, err)
+	require.Equal(t, big.NewInt(5), blk.Number())
+	assert.Equal(t, time2, blk.Time())
+
+	query := ethereum.FilterQuery{
+		FromBlock: big.NewInt(2),
+		ToBlock:   big.NewInt(5),
+		Topics:    [][]common.Hash{topics},
+		Addresses: []common.Address{th.EmitterAddress1, th.EmitterAddress2}}
+
+	gethLogs, err := th.Client.FilterLogs(ctx, query)
+	require.NoError(t, err)
+	require.Len(t, gethLogs, 2)
+
+	lb, _ := th.LogPoller.LatestBlock()
+	th.PollAndSaveLogs(context.Background(), lb+1)
+	lg1, err := th.LogPoller.Logs(0, 20, EmitterABI.Events["Log1"].ID, th.EmitterAddress1)
+	require.NoError(t, err)
+	lg2, err := th.LogPoller.Logs(0, 20, EmitterABI.Events["Log2"].ID, th.EmitterAddress2)
+	require.NoError(t, err)
+
+	// Logs should have correct timestamps
+	b, _ := th.Client.BlockByHash(context.Background(), lg1[0].BlockHash)
+	t.Log(len(lg1), lg1[0].BlockTimestamp)
+	assert.Equal(t, int64(b.Time()), lg1[0].BlockTimestamp.UTC().Unix(), time1)
+	b2, _ := th.Client.BlockByHash(context.Background(), lg2[0].BlockHash)
+	assert.Equal(t, int64(b2.Time()), lg2[0].BlockTimestamp.UTC().Unix(), time2)
 }
 
 func TestLogPoller_SynchronizedWithGeth(t *testing.T) {
@@ -108,7 +395,7 @@ func TestLogPoller_SynchronizedWithGeth(t *testing.T) {
 		t.Log("Starting test", mineOrReorg)
 		chainID := testutils.NewRandomEVMChainID()
 		// Set up a test chain with a log emitting contract deployed.
-		orm := NewORM(chainID, db, lggr, pgtest.NewQConfig(true))
+		orm := logpoller.NewORM(chainID, db, lggr, pgtest.NewQConfig(true))
 		// Note this property test is run concurrently and the sim is not threadsafe.
 		ec := backends.NewSimulatedBackend(map[common.Address]core.GenesisAccount{
 			owner.From: {
@@ -117,12 +404,14 @@ func TestLogPoller_SynchronizedWithGeth(t *testing.T) {
 		}, 10e6)
 		_, _, emitter1, err := log_emitter.DeployLogEmitter(owner, ec)
 		require.NoError(t, err)
-		lp := NewLogPoller(orm, client.NewSimulatedBackendClient(t, ec, chainID), lggr, 15*time.Second, int64(finalityDepth), 3, 2, 1000)
+		lp := logpoller.NewLogPoller(orm, client.NewSimulatedBackendClient(t, ec, chainID), lggr, 15*time.Second, int64(finalityDepth), 3, 2, 1000)
 		for i := 0; i < finalityDepth; i++ { // Have enough blocks that we could reorg the full finalityDepth-1.
 			ec.Commit()
 		}
 		currentBlock := int64(1)
-		currentBlock = lp.PollAndSaveLogs(testutils.Context(t), currentBlock)
+		lp.PollAndSaveLogs(testutils.Context(t), currentBlock)
+		currentBlock, err = lp.LatestBlock()
+		require.NoError(t, err)
 		matchesGeth := func() bool {
 			// Check every block is identical
 			latest, err := ec.BlockByNumber(testutils.Context(t), nil)
@@ -171,7 +460,9 @@ func TestLogPoller_SynchronizedWithGeth(t *testing.T) {
 				require.NoError(t, err)
 				t.Logf("New latest (%v, %x), latest parent %x)\n", latest.NumberU64(), latest.Hash(), latest.ParentHash())
 			}
-			currentBlock = lp.PollAndSaveLogs(testutils.Context(t), currentBlock)
+			lp.PollAndSaveLogs(testutils.Context(t), currentBlock)
+			currentBlock, err = lp.LatestBlock()
+			require.NoError(t, err)
 		}
 		return matchesGeth()
 	}, gen.SliceOfN(numChainInserts, gen.UInt64Range(1, uint64(finalityDepth-1))))) // Max reorg depth is finality depth - 1
@@ -183,7 +474,7 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	th := SetupTH(t, 2, 3, 2)
 
 	// Set up a log poller listening for log emitter logs.
-	err := th.LogPoller.RegisterFilter(Filter{
+	err := th.LogPoller.RegisterFilter(logpoller.Filter{
 		"Test Emitter 1 & 2", []common.Hash{EmitterABI.Events["Log1"].ID, EmitterABI.Events["Log2"].ID},
 		[]common.Address{th.EmitterAddress1, th.EmitterAddress2},
 	})
@@ -197,7 +488,7 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	// Test scenario: single block in chain, no logs.
 	// Chain genesis <- 1
 	// DB: empty
-	newStart := th.LogPoller.PollAndSaveLogs(testutils.Context(t), 1)
+	newStart := th.PollAndSaveLogs(testutils.Context(t), 1)
 	assert.Equal(t, int64(2), newStart)
 
 	// We expect to have saved block 1.
@@ -212,15 +503,15 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	lgs, err := th.ORM.SelectLogsByBlockRange(1, 1)
 	require.NoError(t, err)
 	assert.Equal(t, 0, len(lgs))
-	assertHaveCanonical(t, 1, 1, th.Client, th.ORM)
+	th.assertHaveCanonical(t, 1, 1)
 
 	// Polling again should be a noop, since we are at the latest.
-	newStart = th.LogPoller.PollAndSaveLogs(testutils.Context(t), newStart)
+	newStart = th.PollAndSaveLogs(testutils.Context(t), newStart)
 	assert.Equal(t, int64(2), newStart)
 	latest, err := th.ORM.SelectLatestBlock()
 	require.NoError(t, err)
 	assert.Equal(t, int64(1), latest.BlockNumber)
-	assertHaveCanonical(t, 1, 1, th.Client, th.ORM)
+	th.assertHaveCanonical(t, 1, 1)
 
 	// Test scenario: one log 2 block chain.
 	// Chain gen <- 1 <- 2 (L1)
@@ -230,7 +521,7 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	th.Client.Commit()
 
 	// Polling should get us the L1 log.
-	newStart = th.LogPoller.PollAndSaveLogs(testutils.Context(t), newStart)
+	newStart = th.PollAndSaveLogs(testutils.Context(t), newStart)
 	assert.Equal(t, int64(3), newStart)
 	latest, err = th.ORM.SelectLatestBlock()
 	require.NoError(t, err)
@@ -265,7 +556,7 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	// Create 3 (we need a new block for us to do any polling and detect the reorg).
 	th.Client.Commit()
 
-	newStart = th.LogPoller.PollAndSaveLogs(testutils.Context(t), newStart)
+	newStart = th.PollAndSaveLogs(testutils.Context(t), newStart)
 	assert.Equal(t, int64(4), newStart)
 	latest, err = th.ORM.SelectLatestBlock()
 	require.NoError(t, err)
@@ -274,7 +565,7 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, 1, len(lgs))
 	assert.Equal(t, hexutil.MustDecode(`0x0000000000000000000000000000000000000000000000000000000000000002`), lgs[0].Data)
-	assertHaveCanonical(t, 1, 3, th.Client, th.ORM)
+	th.assertHaveCanonical(t, 1, 3)
 
 	// Test scenario: reorg back to previous tip.
 	// Chain gen <- 1 <- 2 (L1_1) <- 3' (L1_3) <- 4
@@ -286,7 +577,7 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	th.Client.Commit()
 	// Create 4
 	th.Client.Commit()
-	newStart = th.LogPoller.PollAndSaveLogs(testutils.Context(t), newStart)
+	newStart = th.PollAndSaveLogs(testutils.Context(t), newStart)
 	assert.Equal(t, int64(5), newStart)
 	latest, err = th.ORM.SelectLatestBlock()
 	require.NoError(t, err)
@@ -299,9 +590,9 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	assert.Equal(t, hexutil.MustDecode(`0x0000000000000000000000000000000000000000000000000000000000000001`), lgs[0].Data)
 	assert.Equal(t, int64(3), lgs[1].BlockNumber)
 	assert.Equal(t, hexutil.MustDecode(`0x0000000000000000000000000000000000000000000000000000000000000003`), lgs[1].Data)
-	assertHaveCanonical(t, 1, 1, th.Client, th.ORM)
-	assertHaveCanonical(t, 3, 4, th.Client, th.ORM)
-	assertDontHave(t, 2, 2, th.ORM) // 2 gets backfilled
+	th.assertHaveCanonical(t, 1, 1)
+	th.assertHaveCanonical(t, 3, 4)
+	th.assertDontHave(t, 2, 2) // 2 gets backfilled
 
 	// Test scenario: multiple logs per block for many blocks (also after reorg).
 	// Chain gen <- 1 <- 2 (L1_1) <- 3' L1_3 <- 4 <- 5 (L1_4, L2_5) <- 6 (L1_6)
@@ -320,7 +611,7 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	// Create 5
 	th.Client.Commit()
 
-	newStart = th.LogPoller.PollAndSaveLogs(testutils.Context(t), newStart)
+	newStart = th.PollAndSaveLogs(testutils.Context(t), newStart)
 	assert.Equal(t, int64(7), newStart)
 	lgs, err = th.ORM.SelectLogsByBlockRange(4, 6)
 	require.NoError(t, err)
@@ -331,9 +622,9 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	assert.Equal(t, th.EmitterAddress2, lgs[1].Address)
 	assert.Equal(t, hexutil.MustDecode(`0x0000000000000000000000000000000000000000000000000000000000000006`), lgs[2].Data)
 	assert.Equal(t, th.EmitterAddress1, lgs[2].Address)
-	assertHaveCanonical(t, 1, 1, th.Client, th.ORM)
-	assertDontHave(t, 2, 2, th.ORM) // 2 gets backfilled
-	assertHaveCanonical(t, 3, 6, th.Client, th.ORM)
+	th.assertHaveCanonical(t, 1, 1)
+	th.assertDontHave(t, 2, 2) // 2 gets backfilled
+	th.assertHaveCanonical(t, 3, 6)
 
 	// Test scenario: node down for exactly finality + 2 blocks
 	// Note we only backfill up to finalized - 1 blocks, because we need to save the
@@ -347,7 +638,7 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 		require.NoError(t, err)
 		th.Client.Commit()
 	}
-	newStart = th.LogPoller.PollAndSaveLogs(testutils.Context(t), newStart)
+	newStart = th.PollAndSaveLogs(testutils.Context(t), newStart)
 	assert.Equal(t, int64(11), newStart)
 	lgs, err = th.ORM.SelectLogsByBlockRange(7, 9)
 	require.NoError(t, err)
@@ -358,8 +649,8 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	assert.Equal(t, int64(8), lgs[1].BlockNumber)
 	assert.Equal(t, hexutil.MustDecode(`0x0000000000000000000000000000000000000000000000000000000000000009`), lgs[2].Data)
 	assert.Equal(t, int64(9), lgs[2].BlockNumber)
-	assertDontHave(t, 7, 7, th.ORM) // Do not expect to save backfilled blocks.
-	assertHaveCanonical(t, 8, 10, th.Client, th.ORM)
+	th.assertDontHave(t, 7, 7) // Do not expect to save backfilled blocks.
+	th.assertHaveCanonical(t, 8, 10)
 
 	// Test scenario large backfill (multiple batches)
 	// Chain gen <- 1 <- 2 (L1_1) <- 3' L1_3 <- 4 <- 5 (L1_4, L2_5) <- 6 (L1_6) <- 7 (L1_7) <- 8 (L1_8) <- 9 (L1_9) <- 10..16
@@ -373,13 +664,13 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 		require.NoError(t, err)
 		th.Client.Commit()
 	}
-	newStart = th.LogPoller.PollAndSaveLogs(testutils.Context(t), newStart)
+	newStart = th.PollAndSaveLogs(testutils.Context(t), newStart)
 	assert.Equal(t, int64(18), newStart)
 	lgs, err = th.ORM.SelectLogsByBlockRange(11, 17)
 	require.NoError(t, err)
 	assert.Equal(t, 7, len(lgs))
-	assertHaveCanonical(t, 15, 16, th.Client, th.ORM)
-	assertDontHave(t, 11, 14, th.ORM) // Do not expect to save backfilled blocks.
+	th.assertHaveCanonical(t, 15, 16)
+	th.assertDontHave(t, 11, 14) // Do not expect to save backfilled blocks.
 
 	// Verify that a custom block timestamp will get written to db correctly also
 	b, err = th.Client.BlockByNumber(testutils.Context(t), nil)
@@ -394,266 +685,21 @@ func TestLogPoller_PollAndSaveLogs(t *testing.T) {
 	require.Equal(t, uint64(180+time.Hour.Seconds()), b.Time())
 }
 
-func TestLogPoller_Logs(t *testing.T) {
-	t.Parallel()
-	th := SetupTH(t, 2, 3, 2)
-	event1 := EmitterABI.Events["Log1"].ID
-	event2 := EmitterABI.Events["Log2"].ID
-	address1 := common.HexToAddress("0x2ab9a2Dc53736b361b72d900CdF9F78F9406fbbb")
-	address2 := common.HexToAddress("0x6E225058950f237371261C985Db6bDe26df2200E")
-
-	// Block 1-3
-	require.NoError(t, th.ORM.InsertLogs([]Log{
-		GenLog(th.ChainID, 1, 1, "0x3", event1[:], address1),
-		GenLog(th.ChainID, 2, 1, "0x3", event2[:], address2),
-		GenLog(th.ChainID, 1, 2, "0x4", event1[:], address2),
-		GenLog(th.ChainID, 2, 2, "0x4", event2[:], address1),
-		GenLog(th.ChainID, 1, 3, "0x5", event1[:], address1),
-		GenLog(th.ChainID, 2, 3, "0x5", event2[:], address2),
-	}))
-
-	// Select for all Addresses
-	lgs, err := th.ORM.SelectLogsByBlockRange(1, 3)
-	require.NoError(t, err)
-	require.Equal(t, 6, len(lgs))
-	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000003", lgs[0].BlockHash.String())
-	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000003", lgs[1].BlockHash.String())
-	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000004", lgs[2].BlockHash.String())
-	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000004", lgs[3].BlockHash.String())
-	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000005", lgs[4].BlockHash.String())
-	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000005", lgs[5].BlockHash.String())
-
-	// Filter by Address and topic
-	lgs, err = th.ORM.SelectLogsByBlockRangeFilter(1, 3, address1, event1)
-	require.NoError(t, err)
-	require.Equal(t, 2, len(lgs))
-	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000003", lgs[0].BlockHash.String())
-	assert.Equal(t, address1, lgs[0].Address)
-	assert.Equal(t, event1.Bytes(), lgs[0].Topics[0])
-	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000005", lgs[1].BlockHash.String())
-	assert.Equal(t, address1, lgs[1].Address)
-
-	// Filter by block
-	lgs, err = th.ORM.SelectLogsByBlockRangeFilter(2, 2, address2, event1)
-	require.NoError(t, err)
-	require.Equal(t, 1, len(lgs))
-	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000004", lgs[0].BlockHash.String())
-	assert.Equal(t, int64(1), lgs[0].LogIndex)
-	assert.Equal(t, address2, lgs[0].Address)
-	assert.Equal(t, event1.Bytes(), lgs[0].Topics[0])
-}
-
-// Validate that filters stored in log_filters_table match the filters stored in memory
-func validateFiltersTable(t *testing.T, lp *logPoller, orm *ORM) {
-	filters, err := orm.LoadFilters()
-	require.NoError(t, err)
-	require.Equal(t, len(filters), len(lp.filters))
-	for name, dbFilter := range filters {
-		dbFilter := dbFilter
-		memFilter, ok := lp.filters[name]
-		require.True(t, ok)
-		assert.True(t, memFilter.contains(&dbFilter),
-			fmt.Sprintf("in-memory filter %s is missing some addresses or events from db filter table", name))
-		assert.True(t, dbFilter.contains(&memFilter),
-			fmt.Sprintf("db filter table %s is missing some addresses or events from in-memory filter", name))
-	}
-}
-
-func TestLogPoller_RegisterFilter(t *testing.T) {
-	t.Parallel()
-	a1 := common.HexToAddress("0x2ab9a2dc53736b361b72d900cdf9f78f9406fbbb")
-	a2 := common.HexToAddress("0x2ab9a2dc53736b361b72d900cdf9f78f9406fbbc")
-
-	lggr, observedLogs := logger.TestLoggerObserved(t, zapcore.ErrorLevel)
-	chainID := testutils.NewRandomEVMChainID()
-	db := pgtest.NewSqlxDB(t)
-
-	orm := NewORM(chainID, db, lggr, pgtest.NewQConfig(true))
-	lp := NewLogPoller(orm, nil, lggr, 15*time.Second, 1, 1, 2, 1000)
-
-	filter := Filter{"test filter", []common.Hash{EmitterABI.Events["Log1"].ID}, []common.Address{a1}}
-	err := lp.RegisterFilter(filter)
-	require.Error(t, err, "RegisterFilter failed to save filter to db")
-	require.Equal(t, 1, observedLogs.Len())
-	assertForeignConstraintError(t, observedLogs.All()[0], "evm_log_poller_filters", "evm_log_poller_filters_evm_chain_id_fkey")
-
-	db.Close()
-	db = pgtest.NewSqlxDB(t)
-	lggr = logger.TestLogger(t)
-	orm = NewORM(chainID, db, lggr, pgtest.NewQConfig(true))
-
-	// disable check that chain id exists for rest of tests
-	require.NoError(t, utils.JustError(db.Exec(`SET CONSTRAINTS evm_log_poller_filters_evm_chain_id_fkey DEFERRED`)))
-	// Set up a test chain with a log emitting contract deployed.
-
-	lp = NewLogPoller(orm, nil, lggr, 15*time.Second, 1, 1, 2, 1000)
-
-	// We expect a zero filter if nothing registered yet.
-	f := lp.filter(nil, nil, nil)
-	require.Equal(t, 1, len(f.Addresses))
-	assert.Equal(t, common.HexToAddress("0x0000000000000000000000000000000000000000"), f.Addresses[0])
-
-	err = lp.RegisterFilter(Filter{"Emitter Log 1", []common.Hash{EmitterABI.Events["Log1"].ID}, []common.Address{a1}})
-	require.NoError(t, err)
-	assert.Equal(t, []common.Address{a1}, lp.Filter().Addresses)
-	assert.Equal(t, [][]common.Hash{{EmitterABI.Events["Log1"].ID}}, lp.Filter().Topics)
-	validateFiltersTable(t, lp, orm)
-
-	// Should de-dupe EventSigs
-	err = lp.RegisterFilter(Filter{"Emitter Log 1 + 2", []common.Hash{EmitterABI.Events["Log1"].ID, EmitterABI.Events["Log2"].ID}, []common.Address{a2}})
-	require.NoError(t, err)
-	assert.Equal(t, []common.Address{a1, a2}, lp.Filter().Addresses)
-	assert.Equal(t, [][]common.Hash{{EmitterABI.Events["Log1"].ID, EmitterABI.Events["Log2"].ID}}, lp.Filter().Topics)
-	validateFiltersTable(t, lp, orm)
-
-	// Should de-dupe Addresses
-	err = lp.RegisterFilter(Filter{"Emitter Log 1 + 2 dupe", []common.Hash{EmitterABI.Events["Log1"].ID, EmitterABI.Events["Log2"].ID}, []common.Address{a2}})
-	require.NoError(t, err)
-	assert.Equal(t, []common.Address{a1, a2}, lp.Filter().Addresses)
-	assert.Equal(t, [][]common.Hash{{EmitterABI.Events["Log1"].ID, EmitterABI.Events["Log2"].ID}}, lp.Filter().Topics)
-	validateFiltersTable(t, lp, orm)
-
-	// Address required.
-	err = lp.RegisterFilter(Filter{"no address", []common.Hash{EmitterABI.Events["Log1"].ID}, []common.Address{}})
-	require.Error(t, err)
-	// Event required
-	err = lp.RegisterFilter(Filter{"No event", []common.Hash{}, []common.Address{a1}})
-	require.Error(t, err)
-	validateFiltersTable(t, lp, orm)
-
-	// Removing non-existence filter should error.
-	err = lp.UnregisterFilter("filter doesn't exist")
-	require.Error(t, err)
-
-	// Check that all filters are still there
-	_, ok := lp.filters["Emitter Log 1"]
-	require.True(t, ok, "'Emitter Log 1 filter' missing")
-	_, ok = lp.filters["Emitter Log 1 + 2"]
-	require.True(t, ok, "'Emitter Log 1 + 2' filter missing")
-	_, ok = lp.filters["Emitter Log 1 + 2 dupe"]
-	require.True(t, ok, "'Emitter Log 1 + 2 dupe' filter missing")
-
-	// Removing an existing filter should remove it from both memory and db
-	err = lp.UnregisterFilter("Emitter Log 1 + 2")
-	require.NoError(t, err)
-	_, ok = lp.filters["Emitter Log 1 + 2"]
-	require.False(t, ok, "'Emitter Log 1 filter' should have been removed by UnregisterFilter()")
-	require.Len(t, lp.filters, 2)
-	validateFiltersTable(t, lp, orm)
-
-	err = lp.UnregisterFilter("Emitter Log 1 + 2 dupe")
-	require.NoError(t, err)
-	err = lp.UnregisterFilter("Emitter Log 1")
-	require.NoError(t, err)
-	assert.Len(t, lp.filters, 0)
-	filters, err := lp.orm.LoadFilters()
-	require.NoError(t, err)
-	assert.Len(t, filters, 0)
-
-	// Make sure cache was invalidated
-	assert.Len(t, lp.Filter().Addresses, 1)
-	assert.Equal(t, lp.Filter().Addresses[0], common.HexToAddress("0x0000000000000000000000000000000000000000"))
-	assert.Len(t, lp.Filter().Topics, 1)
-	assert.Len(t, lp.Filter().Topics[0], 0)
-}
-
-func assertForeignConstraintError(t *testing.T, observedLog observer.LoggedEntry,
-	table string, constraint string) {
-
-	assert.Equal(t, "SQL ERROR", observedLog.Entry.Message)
-
-	field := observedLog.Context[0]
-	require.Equal(t, zapcore.ErrorType, field.Type)
-	err, ok := field.Interface.(error)
-	var pgErr *pgconn.PgError
-	require.True(t, errors.As(err, &pgErr))
-	require.True(t, ok)
-	assert.Equal(t, "23503", pgErr.SQLState()) // foreign key constraint violation code
-	assert.Equal(t, table, pgErr.TableName)
-	assert.Equal(t, constraint, pgErr.ConstraintName)
-}
-
-func TestLogPoller_DBErrorHandling(t *testing.T) {
-	t.Parallel()
-	lggr, observedLogs := logger.TestLoggerObserved(t, zapcore.WarnLevel)
-	chainID1 := testutils.NewRandomEVMChainID()
-	chainID2 := testutils.NewRandomEVMChainID()
-	db := pgtest.NewSqlxDB(t)
-	o := NewORM(chainID1, db, lggr, pgtest.NewQConfig(true))
-
-	owner := testutils.MustNewSimTransactor(t)
-	ethDB := rawdb.NewMemoryDatabase()
-	ec := backends.NewSimulatedBackendWithDatabase(ethDB, map[common.Address]core.GenesisAccount{
-		owner.From: {
-			Balance: big.NewInt(0).Mul(big.NewInt(10), big.NewInt(1e18)),
-		},
-	}, 10e6)
-	_, _, emitter, err := log_emitter.DeployLogEmitter(owner, ec)
-	require.NoError(t, err)
-	_, err = emitter.EmitLog1(owner, []*big.Int{big.NewInt(9)})
-	require.NoError(t, err)
-	_, err = emitter.EmitLog1(owner, []*big.Int{big.NewInt(7)})
-	require.NoError(t, err)
-	ec.Commit()
-	ec.Commit()
-	ec.Commit()
-
-	lp := NewLogPoller(o, client.NewSimulatedBackendClient(t, ec, chainID2), lggr, 1*time.Hour, 2, 3, 2, 1000)
-	ctx, cancelReplay := context.WithCancel(testutils.Context(t))
-	lp.ctx, lp.cancel = context.WithCancel(testutils.Context(t))
-	defer cancelReplay()
-	defer lp.cancel()
-
-	err = lp.Replay(ctx, 5) // block number too high
-	require.ErrorContains(t, err, "Invalid replay block number")
-
-	// Force a db error while loading the filters (tx aborted, already rolled back)
-	require.Error(t, utils.JustError(db.Exec(`invalid query`)))
-	go func() {
-		err = lp.Replay(ctx, 2)
-		assert.Error(t, err, ErrReplayAbortedByClient)
-	}()
-
-	time.Sleep(100 * time.Millisecond)
-	go lp.run()
-	require.Eventually(t, func() bool {
-		return observedLogs.Len() >= 5
-	}, 2*time.Second, 20*time.Millisecond)
-	lp.cancel()
-	lp.Close()
-	<-lp.done
-
-	logMsgs := make(map[string]int)
-	for _, obs := range observedLogs.All() {
-		_, ok := logMsgs[obs.Entry.Message]
-		if ok {
-			logMsgs[(obs.Entry.Message)] = 1
-		} else {
-			logMsgs[(obs.Entry.Message)]++
-		}
-	}
-
-	assert.Contains(t, logMsgs, "SQL ERROR")
-	assert.Contains(t, logMsgs, "Failed loading filters in main logpoller loop, retrying later")
-	assert.Contains(t, logMsgs, "Error executing replay, could not get fromBlock")
-	assert.Contains(t, logMsgs, "backup log poller ran before filters loaded, skipping")
-}
-
 func TestLogPoller_LoadFilters(t *testing.T) {
 	t.Parallel()
 	th := SetupTH(t, 2, 3, 2)
 
-	filter1 := Filter{"first filter", []common.Hash{
+	filter1 := logpoller.Filter{"first Filter", []common.Hash{
 		EmitterABI.Events["Log1"].ID, EmitterABI.Events["Log2"].ID}, []common.Address{th.EmitterAddress1, th.EmitterAddress2}}
-	filter2 := Filter{"second filter", []common.Hash{
+	filter2 := logpoller.Filter{"second Filter", []common.Hash{
 		EmitterABI.Events["Log2"].ID, EmitterABI.Events["Log3"].ID}, []common.Address{th.EmitterAddress2}}
-	filter3 := Filter{"third filter", []common.Hash{
+	filter3 := logpoller.Filter{"third Filter", []common.Hash{
 		EmitterABI.Events["Log1"].ID}, []common.Address{th.EmitterAddress1, th.EmitterAddress2}}
 
-	assert.True(t, filter1.contains(nil))
-	assert.False(t, filter1.contains(&filter2))
-	assert.False(t, filter2.contains(&filter1))
-	assert.True(t, filter1.contains(&filter3))
+	assert.True(t, filter1.Contains(nil))
+	assert.False(t, filter1.Contains(&filter2))
+	assert.False(t, filter2.Contains(&filter1))
+	assert.True(t, filter1.Contains(&filter3))
 
 	err := th.LogPoller.RegisterFilter(filter1)
 	require.NoError(t, err)
@@ -667,78 +713,27 @@ func TestLogPoller_LoadFilters(t *testing.T) {
 	require.NotNil(t, filters)
 	require.Len(t, filters, 3)
 
-	filter, ok := filters["first filter"]
+	filter, ok := filters["first Filter"]
 	require.True(t, ok)
-	assert.True(t, filter.contains(&filter1))
-	assert.True(t, filter1.contains(&filter))
+	assert.True(t, filter.Contains(&filter1))
+	assert.True(t, filter1.Contains(&filter))
 
-	filter, ok = filters["second filter"]
+	filter, ok = filters["second Filter"]
 	require.True(t, ok)
-	assert.True(t, filter.contains(&filter2))
-	assert.True(t, filter2.contains(&filter))
+	assert.True(t, filter.Contains(&filter2))
+	assert.True(t, filter2.Contains(&filter))
 
-	filter, ok = filters["third filter"]
+	filter, ok = filters["third Filter"]
 	require.True(t, ok)
-	assert.True(t, filter.contains(&filter3))
-	assert.True(t, filter3.contains(&filter))
-}
-
-type ConvertLogsTestCases struct {
-	name     string
-	logs     []types.Log
-	blocks   []LogPollerBlock
-	expected int
-}
-
-func TestLogPoller_ConvertLogs(t *testing.T) {
-	t.Parallel()
-	lggr := logger.TestLogger(t)
-
-	topics := []common.Hash{EmitterABI.Events["Log1"].ID}
-
-	cases := []ConvertLogsTestCases{
-		{"SingleBlock",
-			[]types.Log{{Topics: topics}, {Topics: topics}},
-			[]LogPollerBlock{{BlockTimestamp: time.Now()}},
-			2},
-		{"BlockList",
-			[]types.Log{{Topics: topics}, {Topics: topics}, {Topics: topics}},
-			[]LogPollerBlock{{BlockTimestamp: time.Now()}},
-			3},
-		{"EmptyList",
-			[]types.Log{},
-			[]LogPollerBlock{},
-			0},
-		{"TooManyBlocks",
-			[]types.Log{{}},
-			[]LogPollerBlock{{}, {}},
-			0},
-		{"TooFewBlocks",
-			[]types.Log{{}, {}, {}},
-			[]LogPollerBlock{{}, {}},
-			0},
-	}
-
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			logs := convertLogs(c.logs, c.blocks, lggr, big.NewInt(53))
-			require.Len(t, logs, c.expected)
-			for i := 0; i < c.expected; i++ {
-				if len(c.blocks) == 1 {
-					assert.Equal(t, c.blocks[0].BlockTimestamp, logs[i].BlockTimestamp)
-				} else {
-					assert.Equal(t, logs[i].BlockTimestamp, c.blocks[i].BlockTimestamp)
-				}
-			}
-		})
-	}
+	assert.True(t, filter.Contains(&filter3))
+	assert.True(t, filter3.Contains(&filter))
 }
 
 func TestLogPoller_GetBlocks_Range(t *testing.T) {
 	t.Parallel()
 	th := SetupTH(t, 2, 3, 2)
 
-	err := th.LogPoller.RegisterFilter(Filter{"GetBlocks Test", []common.Hash{
+	err := th.LogPoller.RegisterFilter(logpoller.Filter{"GetBlocks Test", []common.Hash{
 		EmitterABI.Events["Log1"].ID, EmitterABI.Events["Log2"].ID}, []common.Address{th.EmitterAddress1, th.EmitterAddress2}},
 	)
 	require.NoError(t, err)
@@ -842,16 +837,6 @@ func TestLogPoller_GetBlocks_Range(t *testing.T) {
 	cancel()
 	_, err = th.LogPoller.GetBlocksRange(testutils.Context(t), blockNums, qopts)
 	require.NoError(t, err)
-
-	// getBlocksRange returns blocks with a nil client
-	th.LogPoller.ec = nil
-	blockNums = []uint64{1, 2}
-	blocks, err = th.LogPoller.GetBlocksRange(testutils.Context(t), blockNums)
-	require.NoError(t, err)
-	assert.Equal(t, 1, int(blocks[0].BlockNumber))
-	assert.NotEmpty(t, blocks[0].BlockHash)
-	assert.Equal(t, 2, int(blocks[1].BlockNumber))
-	assert.NotEmpty(t, blocks[1].BlockHash)
 }
 
 func TestGetReplayFromBlock(t *testing.T) {
@@ -864,12 +849,12 @@ func TestGetReplayFromBlock(t *testing.T) {
 
 	// Nothing in the DB yet, should use whatever we specify.
 	requested := int64(5)
-	fromBlock, err := th.LogPoller.getReplayFromBlock(testutils.Context(t), requested)
+	fromBlock, err := th.LogPoller.GetReplayFromBlock(testutils.Context(t), requested)
 	require.NoError(t, err)
 	assert.Equal(t, requested, fromBlock)
 
 	// Do a poll, then we should have up to block 11 (blocks 0 & 1 are contract deployments, 2-10 logs).
-	nextBlock := th.LogPoller.PollAndSaveLogs(testutils.Context(t), 1)
+	nextBlock := th.PollAndSaveLogs(testutils.Context(t), 1)
 	require.Equal(t, int64(12), nextBlock)
 
 	// Commit a few more so chain is ahead.
@@ -878,7 +863,7 @@ func TestGetReplayFromBlock(t *testing.T) {
 	}
 	// Should take min(latest, requested), in this case latest.
 	requested = int64(15)
-	fromBlock, err = th.LogPoller.getReplayFromBlock(testutils.Context(t), requested)
+	fromBlock, err = th.LogPoller.GetReplayFromBlock(testutils.Context(t), requested)
 	require.NoError(t, err)
 	latest, err := th.LogPoller.LatestBlock()
 	require.NoError(t, err)
@@ -886,44 +871,7 @@ func TestGetReplayFromBlock(t *testing.T) {
 
 	// Should take min(latest, requested) in this case requested.
 	requested = int64(7)
-	fromBlock, err = th.LogPoller.getReplayFromBlock(testutils.Context(t), requested)
+	fromBlock, err = th.LogPoller.GetReplayFromBlock(testutils.Context(t), requested)
 	require.NoError(t, err)
 	assert.Equal(t, requested, fromBlock)
-}
-
-func TestFilterName(t *testing.T) {
-	t.Parallel()
-	assert.Equal(t, "a - b:c:d", FilterName("a", "b", "c", "d"))
-	assert.Equal(t, "empty args test", FilterName("empty args test"))
-}
-
-func benchmarkFilter(b *testing.B, nFilters, nAddresses, nEvents int) {
-	lggr := logger.TestLogger(b)
-	lp := NewLogPoller(nil, nil, lggr, 1*time.Hour, 2, 3, 2, 1000)
-	for i := 0; i < nFilters; i++ {
-		var addresses []common.Address
-		var events []common.Hash
-		for j := 0; j < nAddresses; j++ {
-			addresses = append(addresses, common.BigToAddress(big.NewInt(int64(j+1))))
-		}
-		for j := 0; j < nEvents; j++ {
-			events = append(events, common.BigToHash(big.NewInt(int64(j+1))))
-		}
-		err := lp.RegisterFilter(Filter{Name: "my filter", EventSigs: events, Addresses: addresses})
-		require.NoError(b, err)
-	}
-	b.ResetTimer()
-	for n := 0; n < b.N; n++ {
-		lp.filter(nil, nil, nil)
-	}
-}
-
-func BenchmarkFilter10_1(b *testing.B) {
-	benchmarkFilter(b, 10, 1, 1)
-}
-func BenchmarkFilter100_10(b *testing.B) {
-	benchmarkFilter(b, 100, 10, 10)
-}
-func BenchmarkFilter1000_100(b *testing.B) {
-	benchmarkFilter(b, 1000, 100, 100)
 }
