+++ core/chains/evm/logpoller/orm_test.go
@@ -1,4 +1,4 @@
-package logpoller
+package logpoller_test
 
 import (
 	"bytes"
@@ -13,8 +13,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/smartcontractkit/chainlink/core/internal/testutils/pgtest"
-	"github.com/smartcontractkit/chainlink/core/logger"
+	"github.com/smartcontractkit/chainlink/core/chains/evm/logpoller"
 	"github.com/smartcontractkit/chainlink/core/utils"
 )
 
@@ -24,20 +23,40 @@ type block struct {
 	timestamp int64
 }
 
-// Setup creates two orms representing logs from different chains.
-func setup(t testing.TB) (*ORM, *ORM) {
-	db := pgtest.NewSqlxDB(t)
-	lggr := logger.TestLogger(t)
-	require.NoError(t, utils.JustError(db.Exec(`SET CONSTRAINTS evm_log_poller_blocks_evm_chain_id_fkey DEFERRED`)))
-	require.NoError(t, utils.JustError(db.Exec(`SET CONSTRAINTS evm_logs_evm_chain_id_fkey DEFERRED`)))
-	o1 := NewORM(big.NewInt(137), db, lggr, pgtest.NewQConfig(true))
-	o2 := NewORM(big.NewInt(138), db, lggr, pgtest.NewQConfig(true))
-	return o1, o2
+func GenLog(chainID *big.Int, logIndex int64, blockNum int64, blockHash string, topic1 []byte, address common.Address) logpoller.Log {
+	return logpoller.Log{
+		EvmChainId:  utils.NewBig(chainID),
+		LogIndex:    logIndex,
+		BlockHash:   common.HexToHash(blockHash),
+		BlockNumber: blockNum,
+		EventSig:    common.BytesToHash(topic1),
+		Topics:      [][]byte{topic1},
+		Address:     address,
+		TxHash:      common.HexToHash("0x1234"),
+		Data:        append([]byte("hello "), byte(blockNum)),
+	}
+}
+
+func TestLogPoller_Batching(t *testing.T) {
+	t.Parallel()
+	th := SetupTH(t, 2, 3, 2)
+	var logs []logpoller.Log
+	// Inserts are limited to 65535 parameters. A log being 10 parameters this results in
+	// a maximum of 6553 log inserts per tx. As inserting more than 6553 would result in
+	// an error without batching, this test makes sure batching is enabled.
+	for i := 0; i < 15000; i++ {
+		logs = append(logs, GenLog(th.ChainID, int64(i+1), 1, "0x3", EmitterABI.Events["Log1"].ID.Bytes(), th.EmitterAddress1))
+	}
+	require.NoError(t, th.ORM.InsertLogs(logs))
+	lgs, err := th.ORM.SelectLogsByBlockRange(1, 1)
+	require.NoError(t, err)
+	// Make sure all logs are inserted
+	require.Equal(t, len(logs), len(lgs))
 }
 
 func TestORM_GetBlocks_From_Range(t *testing.T) {
-
-	o1, _ := setup(t)
+	th := SetupTH(t, 2, 3, 2)
+	o1 := th.ORM
 	// Insert many blocks and read them back together
 	blocks := []block{
 		{
@@ -91,8 +110,8 @@ func TestORM_GetBlocks_From_Range(t *testing.T) {
 }
 
 func TestORM_GetBlocks_From_Range_Recent_Blocks(t *testing.T) {
-
-	o1, _ := setup(t)
+	th := SetupTH(t, 2, 3, 2)
+	o1 := th.ORM
 	// Insert many blocks and read them back together
 	var recentBlocks []block
 	for i := 1; i <= 256; i++ {
@@ -123,14 +142,16 @@ func TestORM_GetBlocks_From_Range_Recent_Blocks(t *testing.T) {
 }
 
 func TestORM(t *testing.T) {
-	o1, o2 := setup(t)
+	th := SetupTH(t, 2, 3, 2)
+	o1 := th.ORM
+	o2 := th.ORM2
 	// Insert and read back a block.
 	require.NoError(t, o1.InsertBlock(common.HexToHash("0x1234"), 10, time.Now()))
 	b, err := o1.SelectBlockByHash(common.HexToHash("0x1234"))
 	require.NoError(t, err)
 	assert.Equal(t, b.BlockNumber, int64(10))
 	assert.Equal(t, b.BlockHash.Bytes(), common.HexToHash("0x1234").Bytes())
-	assert.Equal(t, b.EvmChainId.String(), "137")
+	assert.Equal(t, b.EvmChainId.String(), th.ChainID.String())
 
 	// Insert blocks from a different chain
 	require.NoError(t, o2.InsertBlock(common.HexToHash("0x1234"), 11, time.Now()))
@@ -139,7 +160,7 @@ func TestORM(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, b2.BlockNumber, int64(11))
 	assert.Equal(t, b2.BlockHash.Bytes(), common.HexToHash("0x1234").Bytes())
-	assert.Equal(t, b2.EvmChainId.String(), "138")
+	assert.Equal(t, b2.EvmChainId.String(), th.ChainID2.String())
 
 	latest, err := o1.SelectLatestBlock()
 	require.NoError(t, err)
@@ -168,9 +189,9 @@ func TestORM(t *testing.T) {
 	// Should be able to insert and read back a log.
 	topic := common.HexToHash("0x1599")
 	topic2 := common.HexToHash("0x1600")
-	require.NoError(t, o1.InsertLogs([]Log{
+	require.NoError(t, o1.InsertLogs([]logpoller.Log{
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    1,
 			BlockHash:   common.HexToHash("0x1234"),
 			BlockNumber: int64(10),
@@ -181,7 +202,7 @@ func TestORM(t *testing.T) {
 			Data:        []byte("hello"),
 		},
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    2,
 			BlockHash:   common.HexToHash("0x1234"),
 			BlockNumber: int64(11),
@@ -192,7 +213,7 @@ func TestORM(t *testing.T) {
 			Data:        []byte("hello"),
 		},
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    3,
 			BlockHash:   common.HexToHash("0x1234"),
 			BlockNumber: int64(12),
@@ -203,7 +224,7 @@ func TestORM(t *testing.T) {
 			Data:        []byte("hello"),
 		},
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    4,
 			BlockHash:   common.HexToHash("0x1234"),
 			BlockNumber: int64(13),
@@ -214,7 +235,7 @@ func TestORM(t *testing.T) {
 			Data:        []byte("hello"),
 		},
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    5,
 			BlockHash:   common.HexToHash("0x1234"),
 			BlockNumber: int64(14),
@@ -225,7 +246,7 @@ func TestORM(t *testing.T) {
 			Data:        []byte("hello2"),
 		},
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    6,
 			BlockHash:   common.HexToHash("0x1234"),
 			BlockNumber: int64(15),
@@ -236,7 +257,7 @@ func TestORM(t *testing.T) {
 			Data:        []byte("hello2"),
 		},
 	}))
-	logs, err := o1.selectLogsByBlockRange(10, 10)
+	logs, err := o1.SelectLogsByBlockRange(10, 10)
 	require.NoError(t, err)
 	require.Equal(t, 1, len(logs))
 	assert.Equal(t, []byte("hello"), logs[0].Data)
@@ -316,21 +337,21 @@ func TestORM(t *testing.T) {
 	latest, err = o1.SelectLatestBlock()
 	require.NoError(t, err)
 	t.Log(latest.BlockNumber)
-	logs, err = o1.selectLogsByBlockRange(1, latest.BlockNumber)
+	logs, err = o1.SelectLogsByBlockRange(1, latest.BlockNumber)
 	require.NoError(t, err)
 	require.Equal(t, 0, len(logs))
 }
 
-func insertLogsTopicValueRange(t *testing.T, o *ORM, addr common.Address, blockNumber int, eventSig common.Hash, start, stop int) {
-	var lgs []Log
+func insertLogsTopicValueRange(t *testing.T, chainID *big.Int, o *logpoller.ORM, addr common.Address, blockNumber int, eventSig common.Hash, start, stop int) {
+	var lgs []logpoller.Log
 	for i := start; i <= stop; i++ {
-		lgs = append(lgs, Log{
-			EvmChainId:  utils.NewBig(o.chainID),
+		lgs = append(lgs, logpoller.Log{
+			EvmChainId:  utils.NewBig(chainID),
 			LogIndex:    int64(i),
 			BlockHash:   common.HexToHash("0x1234"),
 			BlockNumber: int64(blockNumber),
 			EventSig:    eventSig,
-			Topics:      [][]byte{eventSig[:], EvmWord(uint64(i)).Bytes()},
+			Topics:      [][]byte{eventSig[:], logpoller.EvmWord(uint64(i)).Bytes()},
 			Address:     addr,
 			TxHash:      common.HexToHash("0x1888"),
 			Data:        []byte("hello"),
@@ -340,54 +361,56 @@ func insertLogsTopicValueRange(t *testing.T, o *ORM, addr common.Address, blockN
 }
 
 func TestORM_IndexedLogs(t *testing.T) {
-	o1, _ := setup(t)
+	th := SetupTH(t, 2, 3, 2)
+	o1 := th.ORM
 	eventSig := common.HexToHash("0x1599")
 	addr := common.HexToAddress("0x1234")
 	require.NoError(t, o1.InsertBlock(common.HexToHash("0x1"), 1, time.Now()))
-	insertLogsTopicValueRange(t, o1, addr, 1, eventSig, 1, 3)
-	insertLogsTopicValueRange(t, o1, addr, 2, eventSig, 4, 4) // unconfirmed
+	insertLogsTopicValueRange(t, th.ChainID, o1, addr, 1, eventSig, 1, 3)
+	insertLogsTopicValueRange(t, th.ChainID, o1, addr, 2, eventSig, 4, 4) // unconfirmed
 
-	lgs, err := o1.SelectIndexedLogs(addr, eventSig, 1, []common.Hash{EvmWord(1)}, 0)
+	lgs, err := o1.SelectIndexedLogs(addr, eventSig, 1, []common.Hash{logpoller.EvmWord(1)}, 0)
 	require.NoError(t, err)
 	require.Equal(t, 1, len(lgs))
-	assert.Equal(t, EvmWord(1).Bytes(), lgs[0].GetTopics()[1].Bytes())
+	assert.Equal(t, logpoller.EvmWord(1).Bytes(), lgs[0].GetTopics()[1].Bytes())
 
-	lgs, err = o1.SelectIndexedLogs(addr, eventSig, 1, []common.Hash{EvmWord(1), EvmWord(2)}, 0)
+	lgs, err = o1.SelectIndexedLogs(addr, eventSig, 1, []common.Hash{logpoller.EvmWord(1), logpoller.EvmWord(2)}, 0)
 	require.NoError(t, err)
 	assert.Equal(t, 2, len(lgs))
 
-	lgs, err = o1.SelectIndexLogsTopicGreaterThan(addr, eventSig, 1, EvmWord(2), 0)
+	lgs, err = o1.SelectIndexLogsTopicGreaterThan(addr, eventSig, 1, logpoller.EvmWord(2), 0)
 	require.NoError(t, err)
 	assert.Equal(t, 2, len(lgs))
 
-	lgs, err = o1.SelectIndexLogsTopicRange(addr, eventSig, 1, EvmWord(3), EvmWord(3), 0)
+	lgs, err = o1.SelectIndexLogsTopicRange(addr, eventSig, 1, logpoller.EvmWord(3), logpoller.EvmWord(3), 0)
 	require.NoError(t, err)
 	assert.Equal(t, 1, len(lgs))
-	assert.Equal(t, EvmWord(3).Bytes(), lgs[0].GetTopics()[1].Bytes())
+	assert.Equal(t, logpoller.EvmWord(3).Bytes(), lgs[0].GetTopics()[1].Bytes())
 
-	lgs, err = o1.SelectIndexLogsTopicRange(addr, eventSig, 1, EvmWord(1), EvmWord(3), 0)
+	lgs, err = o1.SelectIndexLogsTopicRange(addr, eventSig, 1, logpoller.EvmWord(1), logpoller.EvmWord(3), 0)
 	require.NoError(t, err)
 	assert.Equal(t, 3, len(lgs))
 
 	// Check confirmations work as expected.
 	require.NoError(t, o1.InsertBlock(common.HexToHash("0x2"), 2, time.Now()))
-	lgs, err = o1.SelectIndexLogsTopicRange(addr, eventSig, 1, EvmWord(4), EvmWord(4), 1)
+	lgs, err = o1.SelectIndexLogsTopicRange(addr, eventSig, 1, logpoller.EvmWord(4), logpoller.EvmWord(4), 1)
 	require.NoError(t, err)
 	assert.Equal(t, 0, len(lgs))
 	require.NoError(t, o1.InsertBlock(common.HexToHash("0x3"), 3, time.Now()))
-	lgs, err = o1.SelectIndexLogsTopicRange(addr, eventSig, 1, EvmWord(4), EvmWord(4), 1)
+	lgs, err = o1.SelectIndexLogsTopicRange(addr, eventSig, 1, logpoller.EvmWord(4), logpoller.EvmWord(4), 1)
 	require.NoError(t, err)
 	assert.Equal(t, 1, len(lgs))
 }
 
 func TestORM_DataWords(t *testing.T) {
-	o1, _ := setup(t)
+	th := SetupTH(t, 2, 3, 2)
+	o1 := th.ORM
 	eventSig := common.HexToHash("0x1599")
 	addr := common.HexToAddress("0x1234")
 	require.NoError(t, o1.InsertBlock(common.HexToHash("0x1"), 1, time.Now()))
-	require.NoError(t, o1.InsertLogs([]Log{
+	require.NoError(t, o1.InsertLogs([]logpoller.Log{
 		{
-			EvmChainId:  utils.NewBig(o1.chainID),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    int64(0),
 			BlockHash:   common.HexToHash("0x1"),
 			BlockNumber: int64(1),
@@ -395,11 +418,11 @@ func TestORM_DataWords(t *testing.T) {
 			Topics:      [][]byte{eventSig[:]},
 			Address:     addr,
 			TxHash:      common.HexToHash("0x1888"),
-			Data:        EvmWord(1).Bytes(),
+			Data:        logpoller.EvmWord(1).Bytes(),
 		},
 		{
 			// In block 2, unconfirmed to start
-			EvmChainId:  utils.NewBig(o1.chainID),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    int64(1),
 			BlockHash:   common.HexToHash("0x2"),
 			BlockNumber: int64(2),
@@ -407,43 +430,44 @@ func TestORM_DataWords(t *testing.T) {
 			Topics:      [][]byte{eventSig[:]},
 			Address:     addr,
 			TxHash:      common.HexToHash("0x1888"),
-			Data:        append(EvmWord(2).Bytes(), EvmWord(3).Bytes()...),
+			Data:        append(logpoller.EvmWord(2).Bytes(), logpoller.EvmWord(3).Bytes()...),
 		},
 	}))
 	// Outside range should fail.
-	lgs, err := o1.SelectDataWordRange(addr, eventSig, 0, EvmWord(2), EvmWord(2), 0)
+	lgs, err := o1.SelectDataWordRange(addr, eventSig, 0, logpoller.EvmWord(2), logpoller.EvmWord(2), 0)
 	require.NoError(t, err)
 	assert.Equal(t, 0, len(lgs))
 
 	// Range including log should succeed
-	lgs, err = o1.SelectDataWordRange(addr, eventSig, 0, EvmWord(1), EvmWord(2), 0)
+	lgs, err = o1.SelectDataWordRange(addr, eventSig, 0, logpoller.EvmWord(1), logpoller.EvmWord(2), 0)
 	require.NoError(t, err)
 	assert.Equal(t, 1, len(lgs))
 
 	// Range only covering log should succeed
-	lgs, err = o1.SelectDataWordRange(addr, eventSig, 0, EvmWord(1), EvmWord(1), 0)
+	lgs, err = o1.SelectDataWordRange(addr, eventSig, 0, logpoller.EvmWord(1), logpoller.EvmWord(1), 0)
 	require.NoError(t, err)
 	assert.Equal(t, 1, len(lgs))
 
 	// Cannot query for unconfirmed second log.
-	lgs, err = o1.SelectDataWordRange(addr, eventSig, 1, EvmWord(3), EvmWord(3), 0)
+	lgs, err = o1.SelectDataWordRange(addr, eventSig, 1, logpoller.EvmWord(3), logpoller.EvmWord(3), 0)
 	require.NoError(t, err)
 	assert.Equal(t, 0, len(lgs))
 	// Confirm it, then can query.
 	require.NoError(t, o1.InsertBlock(common.HexToHash("0x2"), 2, time.Now()))
-	lgs, err = o1.SelectDataWordRange(addr, eventSig, 1, EvmWord(3), EvmWord(3), 0)
+	lgs, err = o1.SelectDataWordRange(addr, eventSig, 1, logpoller.EvmWord(3), logpoller.EvmWord(3), 0)
 	require.NoError(t, err)
 	assert.Equal(t, 1, len(lgs))
-	assert.Equal(t, lgs[0].Data, append(EvmWord(2).Bytes(), EvmWord(3).Bytes()...))
+	assert.Equal(t, lgs[0].Data, append(logpoller.EvmWord(2).Bytes(), logpoller.EvmWord(3).Bytes()...))
 
 	// Check greater than 1 yields both logs.
-	lgs, err = o1.SelectDataWordGreaterThan(addr, eventSig, 0, EvmWord(1), 0)
+	lgs, err = o1.SelectDataWordGreaterThan(addr, eventSig, 0, logpoller.EvmWord(1), 0)
 	require.NoError(t, err)
 	assert.Equal(t, 2, len(lgs))
 }
 
 func TestORM_SelectLogsWithSigsByBlockRangeFilter(t *testing.T) {
-	o1, _ := setup(t)
+	th := SetupTH(t, 2, 3, 2)
+	o1 := th.ORM
 
 	// Insert logs on different topics, should be able to read them
 	// back using SelectLogsWithSigsByBlockRangeFilter and specifying
@@ -451,9 +475,9 @@ func TestORM_SelectLogsWithSigsByBlockRangeFilter(t *testing.T) {
 	topic := common.HexToHash("0x1599")
 	topic2 := common.HexToHash("0x1600")
 	sourceAddr := common.HexToAddress("0x12345")
-	inputLogs := []Log{
+	inputLogs := []logpoller.Log{
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    1,
 			BlockHash:   common.HexToHash("0x1234"),
 			BlockNumber: int64(10),
@@ -464,7 +488,7 @@ func TestORM_SelectLogsWithSigsByBlockRangeFilter(t *testing.T) {
 			Data:        []byte("hello1"),
 		},
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    2,
 			BlockHash:   common.HexToHash("0x1235"),
 			BlockNumber: int64(11),
@@ -475,7 +499,7 @@ func TestORM_SelectLogsWithSigsByBlockRangeFilter(t *testing.T) {
 			Data:        []byte("hello2"),
 		},
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    3,
 			BlockHash:   common.HexToHash("0x1236"),
 			BlockNumber: int64(12),
@@ -486,7 +510,7 @@ func TestORM_SelectLogsWithSigsByBlockRangeFilter(t *testing.T) {
 			Data:        []byte("hello3"),
 		},
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    4,
 			BlockHash:   common.HexToHash("0x1237"),
 			BlockNumber: int64(13),
@@ -497,7 +521,7 @@ func TestORM_SelectLogsWithSigsByBlockRangeFilter(t *testing.T) {
 			Data:        []byte("hello4"),
 		},
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    5,
 			BlockHash:   common.HexToHash("0x1238"),
 			BlockNumber: int64(14),
@@ -508,7 +532,7 @@ func TestORM_SelectLogsWithSigsByBlockRangeFilter(t *testing.T) {
 			Data:        []byte("hello5"),
 		},
 		{
-			EvmChainId:  utils.NewBigI(137),
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    6,
 			BlockHash:   common.HexToHash("0x1239"),
 			BlockNumber: int64(15),
@@ -536,7 +560,8 @@ func TestORM_SelectLogsWithSigsByBlockRangeFilter(t *testing.T) {
 }
 
 func TestORM_DeleteBlocksBefore(t *testing.T) {
-	o1, _ := setup(t)
+	th := SetupTH(t, 2, 3, 2)
+	o1 := th.ORM
 	require.NoError(t, o1.InsertBlock(common.HexToHash("0x1234"), 1, time.Now()))
 	require.NoError(t, o1.InsertBlock(common.HexToHash("0x1235"), 2, time.Now()))
 	require.NoError(t, o1.DeleteBlocksBefore(1))
@@ -556,13 +581,63 @@ func TestORM_DeleteBlocksBefore(t *testing.T) {
 	require.Equal(t, err, sql.ErrNoRows)
 }
 
+func TestLogPoller_Logs(t *testing.T) {
+	t.Parallel()
+	th := SetupTH(t, 2, 3, 2)
+	event1 := EmitterABI.Events["Log1"].ID
+	event2 := EmitterABI.Events["Log2"].ID
+	address1 := common.HexToAddress("0x2ab9a2Dc53736b361b72d900CdF9F78F9406fbbb")
+	address2 := common.HexToAddress("0x6E225058950f237371261C985Db6bDe26df2200E")
+
+	// Block 1-3
+	require.NoError(t, th.ORM.InsertLogs([]logpoller.Log{
+		GenLog(th.ChainID, 1, 1, "0x3", event1[:], address1),
+		GenLog(th.ChainID, 2, 1, "0x3", event2[:], address2),
+		GenLog(th.ChainID, 1, 2, "0x4", event1[:], address2),
+		GenLog(th.ChainID, 2, 2, "0x4", event2[:], address1),
+		GenLog(th.ChainID, 1, 3, "0x5", event1[:], address1),
+		GenLog(th.ChainID, 2, 3, "0x5", event2[:], address2),
+	}))
+
+	// Select for all Addresses
+	lgs, err := th.ORM.SelectLogsByBlockRange(1, 3)
+	require.NoError(t, err)
+	require.Equal(t, 6, len(lgs))
+	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000003", lgs[0].BlockHash.String())
+	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000003", lgs[1].BlockHash.String())
+	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000004", lgs[2].BlockHash.String())
+	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000004", lgs[3].BlockHash.String())
+	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000005", lgs[4].BlockHash.String())
+	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000005", lgs[5].BlockHash.String())
+
+	// Filter by Address and topic
+	lgs, err = th.ORM.SelectLogsByBlockRangeFilter(1, 3, address1, event1)
+	require.NoError(t, err)
+	require.Equal(t, 2, len(lgs))
+	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000003", lgs[0].BlockHash.String())
+	assert.Equal(t, address1, lgs[0].Address)
+	assert.Equal(t, event1.Bytes(), lgs[0].Topics[0])
+	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000005", lgs[1].BlockHash.String())
+	assert.Equal(t, address1, lgs[1].Address)
+
+	// Filter by block
+	lgs, err = th.ORM.SelectLogsByBlockRangeFilter(2, 2, address2, event1)
+	require.NoError(t, err)
+	require.Equal(t, 1, len(lgs))
+	assert.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000004", lgs[0].BlockHash.String())
+	assert.Equal(t, int64(1), lgs[0].LogIndex)
+	assert.Equal(t, address2, lgs[0].Address)
+	assert.Equal(t, event1.Bytes(), lgs[0].Topics[0])
+}
+
 func BenchmarkLogs(b *testing.B) {
-	o, _ := setup(b)
-	var lgs []Log
+	th := SetupTH(b, 2, 3, 2)
+	o := th.ORM
+	var lgs []logpoller.Log
 	addr := common.HexToAddress("0x1234")
 	for i := 0; i < 10_000; i++ {
-		lgs = append(lgs, Log{
-			EvmChainId:  utils.NewBig(o.chainID),
+		lgs = append(lgs, logpoller.Log{
+			EvmChainId:  utils.NewBig(th.ChainID),
 			LogIndex:    int64(i),
 			BlockHash:   common.HexToHash("0x1"),
 			BlockNumber: 1,
@@ -576,7 +651,7 @@ func BenchmarkLogs(b *testing.B) {
 	require.NoError(b, o.InsertLogs(lgs))
 	b.ResetTimer()
 	for n := 0; n < b.N; n++ {
-		_, err := o.SelectDataWordRange(addr, EmitterABI.Events["Log1"].ID, 0, EvmWord(8000), EvmWord(8002), 0)
+		_, err := o.SelectDataWordRange(addr, EmitterABI.Events["Log1"].ID, 0, logpoller.EvmWord(8000), logpoller.EvmWord(8002), 0)
 		require.NoError(b, err)
 	}
 }
