+++ core/internal/testutils/testutils.go
@@ -46,7 +46,7 @@ var SimulatedChainID = big.NewInt(1337)
 
 // MustNewSimTransactor returns a transactor for interacting with the
 // geth simulated backend.
-func MustNewSimTransactor(t *testing.T) *bind.TransactOpts {
+func MustNewSimTransactor(t testing.TB) *bind.TransactOpts {
 	key, err := crypto.GenerateKey()
 	require.NoError(t, err)
 	transactor, err := bind.NewKeyedTransactorWithChainID(key, SimulatedChainID)
